Mar 30, 2024, 3:01 PM

+++ Item.cpp
@@ -30,7 +30,7 @@ int Item::getCostToShip() const
 {
   return costToShip;
 }
-int Item::getReward() const
+double Item::getReward() const
 {
   return reward;
 }
